Add %optional tag
Add the %optional tag from rpm5 -- it allows to mark a file in the file
list as optional, meaning if it's there, it's packaged, but if it isn't
there, it will be ignored.

A typical use case is
%optional %{_includedir}/arch-specific-file.h
(easier and shorter than using %ifarch and having to list all
architectures that have the file -- will also automatically do the right
thing on a new architecture being added)

Another typical use case is in autogenerated file lists, e.g.
per-language subpackages in the likes of libreoffice where all languages
include translations, some languages additionally include dictionaries,
etc. -- with %optional, the same file list can be used for every
language subpackage.

Signed-off-by: Bernhard Rosenkränzer <bero@lindev.ch>