ROSA: fix KF5_HTML directory
If kf5-config is not installed, use hardcoded value of KF5_HTML,
otherwise query kf5-config.

kf5-config belongs to package kdelibs4support which is not installed
in a typical KF5 build environment.
BuildRequiring it just for one variable does not make sense.
That variable was hardcoded in RPM 5.
It is not a good practice, but I do not see any better solutions.

[root@rosa-2019 ~]# kf5-config --expandvars --install html
/usr/share/doc5/HTML/